Clippy

+++ src/ringbuffer.rs
@@ -107,9 +107,9 @@ pub mod tests {
     #[test]
     fn test_peek() {
         let mut rb = RingBuffer::<f32>::new(3);
-        rb.add(1 as f32);
-        rb.add(2 as f32);
-        rb.add(3 as f32);
+        rb.add(1_f32);
+        rb.add(2_f32);
+        rb.add(3_f32);
 
         assert!(rb.peek(0) == &1f32);
         assert!(rb.peek(1) == &2f32);
@@ -120,17 +120,17 @@ pub mod tests {
     fn test_use() {
         let mut rb = RingBuffer::<f32>::new(3);
 
-        rb.add(1 as f32);
-        rb.add(2 as f32);
-        rb.add(3 as f32);
+        rb.add(1_f32);
+        rb.add(2_f32);
+        rb.add(3_f32);
 
         assert!(rb.peek(0) == &1f32);
         assert!(rb.peek(1) == &2f32);
         assert!(rb.peek(2) == &3f32);
 
-        rb.add(4 as f32);
-        rb.add(5 as f32);
-        rb.add(6 as f32);
+        rb.add(4_f32);
+        rb.add(5_f32);
+        rb.add(6_f32);
 
         assert!(rb.peek(0) == &4f32);
         assert!(rb.peek(1) == &5f32);
